lab 5

+++ TPIS/FormJournalOperation.cs
@@ -68,6 +68,20 @@ namespace TPIS
             selectTable(ConnectionString, selectCommand);
         }
 
+        public object selectValue(string ConnectionString, String selectCommand)
+        {
+            SQLiteConnection connect = new SQLiteConnection(ConnectionString);
+            connect.Open();
+            SQLiteCommand command = new SQLiteCommand(selectCommand, connect);
+            SQLiteDataReader reader = command.ExecuteReader();
+            object value = "";
+            while (reader.Read())
+            {
+                value = reader[0];
+            }
+            connect.Close(); return value;
+        }
+
         private void buttonUpdate_Click(object sender, EventArgs e)
         {
           //  string ConnectionString = @"Data Source=" + sPath + ";New=False;Version=3";
@@ -81,11 +95,79 @@ namespace TPIS
 
         private void buttonDel_Click(object sender, EventArgs e)
         {
-           // string ConnectionString = @"Data Source=" + sPath + ";New=False;Version=3";
+            int countMaterialDebet = 0;
+            int countMaterialKredit = 0;
+            string Subkonto1Dt = "";
+            string Subkonto2Dt = "";
+            string Subkonto1Kt = "";
+            string Subkonto2Kt = "";
+            string typeoperation = "";
+
+            // string ConnectionString = @"Data Source=" + sPath + ";New=False;Version=3";
             int CurrentRow = dataGridView1.SelectedCells[0].RowIndex;
             string valueId = dataGridView1[0, CurrentRow].Value.ToString();
-            String deleteJournalOperation = "delete from Journal_Operation where idOperation=" + valueId;
-            changeValue(ConnectionString, deleteJournalOperation);
+
+            String selectSubkonto1Dt = "Select Subkonto1Dt from Journal_Provodok where JournalOperationKod=" + valueId;
+            object subkonto1DtObject = selectValue(ConnectionString, selectSubkonto1Dt);
+            if (Convert.ToString(subkonto1DtObject) == "") subkonto1DtObject = 0;
+            Subkonto1Dt = subkonto1DtObject.ToString();
+            String selectSubkonto2Dt = "Select Subkonto2Dt from Journal_Provodok where JournalOperationKod=" + valueId;
+            object subkonto2DtObject = selectValue(ConnectionString, selectSubkonto2Dt);
+            if (Convert.ToString(subkonto2DtObject) == "") subkonto2DtObject = 0;
+            Subkonto2Dt = subkonto2DtObject.ToString();
+            String selectSubkonto1Kt = "Select Subkonto1Kt from Journal_Provodok where JournalOperationKod=" + valueId;
+            object subkonto1KtObject = selectValue(ConnectionString, selectSubkonto1Kt);
+            if (Convert.ToString(subkonto1KtObject) == "") subkonto1KtObject = 0;
+            Subkonto1Kt = subkonto1KtObject.ToString();
+            String selectSubkonto2Kt = "Select Subkonto2Kt from Journal_Provodok where JournalOperationKod=" + valueId;
+            object subkonto2KtObject = selectValue(ConnectionString, selectSubkonto2Kt);
+            if (Convert.ToString(subkonto2KtObject) == "") subkonto2KtObject = 0;
+            Subkonto2Kt = subkonto2KtObject.ToString();
+            String selectTypeOperation = "Select Type_operation from Journal_Operation where idOperation=" + valueId;
+            object typeOperationObject = selectValue(ConnectionString, selectTypeOperation);
+            if (Convert.ToString(typeOperationObject) == "") typeOperationObject = "";
+            typeoperation = typeOperationObject.ToString();
+
+            String selectCount = "Select Count from Journal_Provodok where JournalOperationKod=" + valueId;
+            object countObject = selectValue(ConnectionString, selectCount);
+            if (Convert.ToString(countObject) == "") countObject = 0;
+            int count = Convert.ToInt32(countObject);
+
+            //подсчет количества материала по дебету
+            String selectCountDebet = "select sum(Count) from Journal_Provodok where Subkonto1Dt LIKE '%" +Subkonto1Dt + "%' and Subkonto2Dt LIKE '%" + Subkonto2Dt + "%'";
+            object countMaterialDebetObject = selectValue(ConnectionString, selectCountDebet);
+            if (Convert.ToString(countMaterialDebetObject) == "") countMaterialDebetObject = 0;
+            countMaterialDebet = Convert.ToInt32(countMaterialDebetObject);
+
+            //подсчет количества материала по кредиту
+            String selectCountKredit = "select sum(Count) from Journal_Provodok where Subkonto1Kt LIKE '%" + Subkonto1Dt + "%' and Subkonto2Kt LIKE '%" + Subkonto2Dt + "%'";
+            object countMaterialKreditObject = selectValue(ConnectionString, selectCountKredit);
+            if (Convert.ToString(countMaterialKreditObject) == "") countMaterialKreditObject = 0;
+            countMaterialKredit = Convert.ToInt32(countMaterialKreditObject);
+
+            int differenceMaterialCount = countMaterialDebet - countMaterialKredit - count;
+
+            if (typeoperation == "Поступление материалов")
+            {
+                if (differenceMaterialCount >= 0)
+                {
+                    String deleteJournalOperation = "delete from Journal_Operation where idOperation=" + valueId;
+                    String deleteJournalProvodok = "delete from Journal_Provodok where JournalOperationKod=" + valueId;
+                    changeValue(ConnectionString, deleteJournalOperation);
+                    changeValue(ConnectionString, deleteJournalProvodok);
+                }
+                else
+                {
+                    MessageBox.Show("Невозможно удалить операцию, удалите операции после данной ", "Сообщение", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+            }
+            else
+            {
+                String deleteJournalOperation = "delete from Journal_Operation where idOperation=" + valueId;
+                String deleteJournalProvodok = "delete from Journal_Provodok where JournalOperationKod=" + valueId;
+                changeValue(ConnectionString, deleteJournalOperation);
+                changeValue(ConnectionString, deleteJournalProvodok);
+            }
 
             String selectCommand1 = "Select idOperation, Type_operation, Sum_operation, Date_operation, Comment from Journal_Operation";
             selectTable(ConnectionString, selectCommand1);
